src/backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
supabase>=2.11.0  # Updated to support new API keys (sb_secret_*, sb_publishable_*)
# psycopg2-binary==2.9.9  # Removed - not needed with Supabase client
# sqlalchemy==2.0.25  # Removed - not needed with Supabase client

# AI Providers
openai==1.10.0
anthropic==0.8.1
google-generativeai==0.3.2
httpx>=0.26  # Updated for new Supabase compatibility

# Email
resend==0.7.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Utilities
pydantic>=2.10.0  # Use latest with pre-built binaries
pydantic-settings>=2.6.0
aiofiles==23.2.1

# File processing
PyPDF2==3.0.1
python-docx==1.1.0
reportlab==4.0.7
nest-asyncio==1.6.0  # For nested event loops (needed for PDF.co async fallback)
# pillow==10.2.0  # Removed - not critical for MVP, requires compilation

# Validation and data
email-validator==2.1.0
phonenumbers==8.13.27

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
flake8==7.0.0
mypy==1.8.0

